Abstracted out common code into ScriptChaosType
